+++ spec/protovalidate_spec.rb
@@ -24,6 +24,14 @@ RSpec.describe Protovalidate do
     end
   end
 
+  describe ".register_all" do
+    it "registers with the shared validator" do
+      allow(Protovalidate.validator).to receive(:register_all)
+      Protovalidate.register_all
+      expect(Protovalidate.validator).to have_received(:register_all)
+    end
+  end
+
   describe ".validate" do
     it "returns nil for a valid message" do
       expect(Protovalidate.validate(valid_user)).to be_nil